requirements.txt
requests
django
psycopg
opencv-python
matplotlib
Pillow
pysubs2
webvtt-py
moviepy
